[Refactor] 클라이언트 전용 풀페이지 컴포넌트 분리
