+++ src/app/page.tsx
@@ -1,52 +1,11 @@
-"use client";
+import FullpageContainer from "@/components/Home/FullpageContainer";
 
-import { useState } from "react";
-import { SectionsContainer, Section } from "react-fullpage";
-import MainSection from "@/components/Home/MainSection";
-import SubSection01 from "@/components/Home/SubSection01";
-import SubSection02 from "@/components/Home/SubSection02";
-import SubSection03 from "@/components/Home/SubSection03";
-import SubSection04 from "@/components/Home/SubSection04";
-
-type ScrollEvent = {
-  activeSection: number;
-};
-
-export default function Home() {
-  const [initialActiveSection, setInitialActiveSection] = useState<number | null>(null);
-  const onScroll = (page: ScrollEvent) => {
-    if (initialActiveSection === null) setInitialActiveSection(page.activeSection);
-  };
-
-  let options = {
-    scrollCallback: onScroll,
-    sectionClassName: "section",
-    anchors: ["MainSection", "SubSection01", "SubSection02", "SubSection03", "SubSection04"],
-    scrollBar: false,
-    navigation: true,
-    verticalAlign: false,
-    arrowNavigation: true,
-  };
-
+const Home = () => {
   return (
     <main className="relative">
-      <SectionsContainer {...options} activeSection={initialActiveSection}>
-        <Section>
-          <MainSection />
-        </Section>
-        <Section>
-          <SubSection01 />
-        </Section>
-        <Section>
-          <SubSection02 />
-        </Section>
-        <Section>
-          <SubSection03 />
-        </Section>
-        <Section>
-          <SubSection04 />
-        </Section>
-      </SectionsContainer>
+      <FullpageContainer />
     </main>
   );
-}
+};
+
+export default Home;
