+++ src/components/Home/FullpageContainer.tsx
@@ -1,0 +1,53 @@
+"use client";
+
+import { useState } from "react";
+import { SectionsContainer, Section } from "react-fullpage";
+import MainSection from "@/components/Home/MainSection";
+import SubSection01 from "@/components/Home/SubSection01";
+import SubSection02 from "@/components/Home/SubSection02";
+import SubSection03 from "@/components/Home/SubSection03";
+import SubSection04 from "@/components/Home/SubSection04";
+
+type ScrollEvent = {
+  activeSection: number;
+};
+
+const FullpageContainer = () => {
+  const [initialActiveSection, setInitialActiveSection] = useState<number | null>(null);
+
+  const onScroll = (page: ScrollEvent) => {
+    if (initialActiveSection === null) setInitialActiveSection(page.activeSection);
+  };
+
+  const options = {
+    scrollCallback: onScroll,
+    sectionClassName: "section",
+    anchors: ["MainSection", "SubSection01", "SubSection02", "SubSection03", "SubSection04"],
+    scrollBar: false,
+    navigation: true,
+    verticalAlign: false,
+    arrowNavigation: true,
+  };
+
+  return (
+    <SectionsContainer {...options} activeSection={initialActiveSection}>
+      <Section>
+        <MainSection />
+      </Section>
+      <Section>
+        <SubSection01 />
+      </Section>
+      <Section>
+        <SubSection02 />
+      </Section>
+      <Section>
+        <SubSection03 />
+      </Section>
+      <Section>
+        <SubSection04 />
+      </Section>
+    </SectionsContainer>
+  );
+};
+
+export default FullpageContainer;
